% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci_binom}
\alias{ci_binom}
\title{Proporcijos PI: 2 grupės}
\usage{
ci_binom(x, n, method = "modified wilson", conf.level = 0.95, ...)
}
\arguments{
\item{x}{Mus dominančių
\emph{arba} mums palankių įvykių skaičius
\emph{arba} mus dominančios grupės dydis.}

\item{n}{Įvykių skaičius iš viso. / Imties dydis.}

\item{method}{Skaičiavimo metodas (\code{"modified wilson"}, \code{"wilson"},
\code{"agresti-coull"} ir kiti variantai, aprašyti
\code{\link[DescTools:BinomCI]{DescTools::BinomCI()}} dokumentacijoje).}

\item{conf.level}{Pasikliovimo lygmuo. Numatytoji reikšmė – 0.95.}

\item{...}{Kiti parametrai, kuriuos priima
\code{\link[DescTools:BinomCI]{DescTools::BinomCI()}}. Žiūrėti šios funkcijos dokumentaciją.}
}
\value{
Rezultatas – duomenų lentelė, kurios stulpeliai:
\itemize{
\item \code{est} (\verb{<dbl>}) – proporcijos įvertis,
\item \code{lwr.ci}, \code{upr.ci}  (\verb{<dbl>}) – (lower CI, upper CI)
apatinė ir viršutinė proporcijos pasikliautinojo intervalo ribos.
\item \code{x} (\verb{<int>}) – Mus dominančių įvykių skaičius / Mus dominančios grupės dydis.
\item \code{n} (\verb{<int>}) – Įvykių skaičius iš viso. / Imties dydis.
}
}
\description{
Dvireikšmio kintamojo mus dominančios reikšmės proporcijos pasikliautinojo
intervalo (PI) skaičiavimo funkcija, kuri yra patobulintas
\code{\link[DescTools:BinomCI]{DescTools::BinomCI()}} variantas. Rezultatas – duomenų lentelė.
}
\details{
Ši funkcija naudojama taip pat, kaip \code{\link[DescTools:BinomCI]{DescTools::BinomCI()}}, tik numatytasis
metodas yra modifikuotasis Wilson metodas, o rezultatas – duomenų lentelė,
o ne vektorius. Todėl, pvz., rezultatą galima braižyti naudojant
\pkg{ggplot2}.
}
\examples{
x <- 54  # mus dominančių įvykių skaičius
n <- 80  # įvykių skaičius iš viso
ci_binom(x = 54, n = 80)

# Simuliacija su skirtingais imties dydžiais
ci_binom(x = 54, n = c(80, 100, 512))

# PI skaičiavimas kiekvienai grupei atskirai
y = c(23, 45)
ci_binom(y, n = sum(y))


}
