% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci_mean_t}
\alias{ci_mean_t}
\title{Vidurkio PI, suskaičiuotas pagal duomenis}
\usage{
ci_mean_t(.data, x, conf.level = 0.95, ...)
}
\arguments{
\item{.data}{Duomenų lentelė.}

\item{x}{Stulpelio pavadinimas (be kabučių).}

\item{conf.level}{Pasikliovimo lygmuo. Numatytoji reikšmė – 0.95.}

\item{...}{Kiti parametrai, kuriuos priima \code{\link[DescTools:MeanCI]{DescTools::MeanCI()}}.
Žiūrėti šios funkcijos dokumentaciją.}
}
\value{
Rezultatas – duomenų lentelė, kurioje yra šie stulpeliai:
\itemize{
\item (jei yra) grupavimo kintamųjų pavadinimai;
\item \code{mean} (\verb{<dbl>}) – vidurkio įvertis;
\item \code{lwr.ci}, \code{upr.ci} (\verb{<dbl>}) – (lower CI, upper CI) apatinė ir viršutinė
pasikliautinojo intervalo ribos.
}
}
\description{
Funkcija \code{ci_mean_t()} skaičiuoja vidurkio pasikliautinajį intervalą (PI)
pagal \emph{klasikinę} formulę su t (Stjudento) koeficientu, kai duomenys
pateikti \emph{duomenų lentelės pavidalu}. Ši funkcija yra patobulinta
\code{\link[DescTools:MeanCI]{DescTools::MeanCI()}}, reaguojanti į \code{\link[dplyr:group_by]{dplyr::group_by()}}, tad skaičiavimus
gali atlikti ir pogrupiams. Rezultatas – duomenų lentelė.
}
\examples{
# Pavyzdžiai
data(npk, package = "datasets")
head(npk)

# Kintamojo `yield` vidurkio PI skaičiavimas
ci_mean_t(npk, yield)

# PI skaičiavimas naudojant jungimo operatorių
npk |> ci_mean_t(yield)

# PI skaičiavimas grupuojant pagal vieną kintamąjį
npk |> dplyr::group_by(N) |> ci_mean_t(yield)

# PI skaičiavimas grupuojant pagal 3 kintamuosius
npk |> dplyr::group_by(N, P, K) |> ci_mean_t(yield)

}
