% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci_multinom}
\alias{ci_multinom}
\title{Proporcijos PI: 3 ar daugiau grupių}
\usage{
ci_multinom(
  x,
  method = "goodman",
  conf.level = 0.95,
  gr_colname = "group",
  ...
)
}
\arguments{
\item{x}{Vektorius su grupių dydžiais.
Geriausia, jei vektoriaus elementai turėtų prasmingu pavadinimus
(žiūrėti pavyzdžius).}

\item{method}{Skaičiavimo metodas (\code{"goodman"}, \code{"sisonglaz"}, \code{"cplus1"}
ir kiti variantai, aprašyti \code{\link[DescTools:MultinomCI]{DescTools::MultinomCI()}} dokumentacijoje).}

\item{conf.level}{Pasikliovimo lygmuo. Numatytoji reikšmė – 0.95.}

\item{gr_colname}{Stulpelio pavadinimas (kabutėse), kuriame bus
parašyti grupių pavadinimai. Numatytoji reikšmė yra \code{"group"}.}

\item{...}{Kiti parametrai, kuriuos priima \code{\link[DescTools:MultinomCI]{DescTools::MultinomCI()}}.
Žiūrėti šios funkcijos dokumentaciją.}
}
\value{
Rezultatas – duomenų lentelė, kurios stulpeliai:
\itemize{
\item \code{group} arba kitas vartotojo pasirinktas pavadinimas stulpeliui su grupių
pavadinimams, numatytoji reikšmė
(\verb{<fct>}).
\item \code{est} (\verb{<dbl>}) – proporcijos įvertis.
\item \code{lwr.ci}, \code{upr.ci} (\verb{<dbl>}) – (lower CI, upper CI) apatinė ir viršutinė
proporcijos pasikliautinojo intervalo ribos.
\item \code{x} (\verb{<int>}) – Grupės dydis.
\item \code{n} (\verb{<int>}) – Įvykių skaičius iš viso. / Imties dydis.
}
}
\description{
Daugiareikšmio (k ≥ 3) kintamojo reikšmių proporcijų \emph{vienu metu} skaičiuojamų
pasikliautinųjų intervalų (PI) skaičiavimo funkcija, kuri yra patobulintas
\code{\link[DescTools:MultinomCI]{DescTools::MultinomCI()}} variantas. Rezultatas – duomenų lentelė.
}
\details{
Ši funkcija naudojama taip pat, kaip \code{\link[DescTools:MultinomCI]{DescTools::MultinomCI()}}, tik
numatytasis metodas yra Goodman metodas, o rezultatas – duomenų lentelė,
o ne vektorius.
Todėl rezultatą galima patogiai braižyti naudojant \pkg{ggplot2}.
}
\examples{
# Dažniai be pavadinimų
ci_multinom(c(20, 35, 54))

# Nurodytas skaičiavimo metodas
ci_multinom(c(20, 35, 54), method = "goodman")

# Dažniai su grupių pavadinimais
x <- c("dideli" = 20, "vidutiniai" = 35, "maži" = 54)
ci_multinom(x, method = "goodman")

# Dažniai su grupių pavadinimais ir jungimo operatorius
c("dideli" = 20, "vidutiniai" = 35, "maži" = 54) |>
  ci_multinom()

# Kitas metodas
c("dideli" = 33, "vidutiniai" = 35, "maži" = 30) |>
  ci_multinom(method = "sisonglaz")


}
