\name{ETCiwd}
\alias{ETCiwd}
\title{Title ETC}
\usage{ETOiwd(data.eto, farms, rastera, date.start, date.end=date.start, stat.coord)}
\arguments{
\item{data.eto}{Comment.}
\item{farms}{Comment.}
\item{rastera}{Comment.}
\item{date.start}{Comment.}
\item{date.end}{Comment.}
\item{stat.coord}{Comment.}
}
\value{
A list containing:

\item{eto.mean.farms}{Comment.}
\item{idw.eto}{Comment.}
\item{eto.farms}{Comment.}
}
\description{
Prints 'Hello, world!'.
}
\author{
Francisco J. Rodriguez-Cortes <frrodriguezc@unal.edu.co>, Fabio H. Sepulveda-Murillo
}
\references{
Hodam, S., Sarkar, S., Marak, A. G., Bandyopadhyay, A., & Bhadra, A. (2017). Spatial interpolation of reference evapotranspiration in India: comparison of IDW and Kriging Methods. Journal of The Institution of Engineers (India): Series A, 98(4), 511-524.
}
\examples{

}
