\name{PI-package}
\alias{PI-package}
\alias{PI}
\docType{package}
\title{
\packageTitle{PI}
}
\description{
\packageDescription{PI}
}
\details{
\tabular{ll}{
Package: \tab PI\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-05-01\cr
License: \tab GPL (>= 2)\cr
}
}

