\name{plot.pi}
\alias{plot.pi}
\title{Hello, World!}
\usage{plot.pi(X,...)}
\arguments{
\item{X}{Comment.}
\item{...}{Comment.}
}
\description{
Prints 'Hello, world!'.
}
\author{
Francisco J. Rodriguez-Cortes <frrodriguezc@unal.edu.co>, Fabio H. Sepulveda-Murillo
}
\references{
Hodam, S., Sarkar, S., Marak, A. G., Bandyopadhyay, A., & Bhadra, A. (2017). Spatial interpolation of reference evapotranspiration in India: comparison of IDW and Kriging Methods. Journal of The Institution of Engineers (India): Series A, 98(4), 511-524.
}
\examples{

}
