% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_coin_pi.R
\name{print.coin_pi_estimate}
\alias{print.coin_pi_estimate}
\title{Print a \code{coin_pi_estimate}}
\usage{
\method{print}{coin_pi_estimate}(x, ...)
}
\arguments{
\item{x}{A \code{coin_pi_estimate} object from \code{\link[=simulate_coin_pi]{simulate_coin_pi()}}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a \code{coin_pi_estimate}
}
